Moving AtlasDbFactory to api and annotating included factories
