+++ atlasdb-api/build.gradle
@@ -2,6 +2,7 @@ apply from: "../gradle/shared.gradle"
 
 dependencies {
   compile(project(":atlasdb-commons"))
+  compile(project(":timestamp-api"))
   compile 'javax.ws.rs:javax.ws.rs-api:2.0.1'
   compile 'com.fasterxml.jackson.core:jackson-annotations:' + libVersions.jackson_annotation
 }
