+++ atlasdb-api/src/main/java/com/palantir/atlasdb/factory/api/AtlasDbFactory.java
@@ -13,7 +13,7 @@
  * See the License for the specific language governing permissions and
  * limitations under the License.
  */
-package com.palantir.atlasdb.spi;
+package com.palantir.atlasdb.factory.api;
 
 import com.palantir.atlasdb.keyvalue.api.KeyValueService;
 import com.palantir.atlasdb.keyvalue.api.KeyValueServiceConfig;
