+++ atlasdb-impl-shared/src/main/java/com/palantir/atlasdb/keyvalue/partition/PartitionedAtlasDbFactory.java
@@ -16,8 +16,8 @@
 package com.palantir.atlasdb.keyvalue.partition;
 
 import com.google.auto.service.AutoService;
+import com.palantir.atlasdb.factory.api.AtlasDbFactory;
 import com.palantir.atlasdb.keyvalue.api.KeyValueService;
-import com.palantir.atlasdb.spi.AtlasDbFactory;
 import com.palantir.atlasdb.keyvalue.api.KeyValueServiceConfig;
 import com.palantir.timestamp.PersistentTimestampService;
 import com.palantir.timestamp.TimestampService;
