+++ atlasdb-impl-shared/src/main/java/com/palantir/atlasdb/keyvalue/partition/PartitionedKeyValueConfiguration.java
@@ -21,8 +21,8 @@ import org.immutables.value.Value;
 
 import com.fasterxml.jackson.annotation.JsonTypeName;
 import com.google.auto.service.AutoService;
-import com.palantir.atlasdb.keyvalue.partition.quorum.QuorumParameters;
 import com.palantir.atlasdb.keyvalue.api.KeyValueServiceConfig;
+import com.palantir.atlasdb.keyvalue.partition.quorum.QuorumParameters;
 
 /**
  * This class is to provide configuration for PartitionedKeyValueService.
