+++ atlasdb-jdbc/build.gradle
@@ -4,7 +4,6 @@ apply from: "../gradle/shared.gradle"
 
 dependencies {
   compile project(":atlasdb-client")
-  compile project(":atlasdb-spi")
   compile project(":timestamp-impl")
 
   compile "io.dropwizard:dropwizard-jackson:" + libVersions.dropwizard
