+++ atlasdb-rocksdb/build.gradle
@@ -4,11 +4,11 @@ apply from: "../gradle/shared.gradle"
 
 dependencies {
   compile project(":atlasdb-client")
-  compile project(":atlasdb-spi")
   compile project(':timestamp-impl')
 
   compile "org.rocksdb:rocksdbjni:4.1.0"
 
   processor 'org.immutables:value:2.0.21'
+  processor "com.google.auto.service:auto-service:1.0-rc2"
 }
